Refactor core into modular mixins
